requirements.txt
Flask
pandas
numpy
groq
python-dotenv
pytesseract
Pillow
